Create Super market billing system
In this project we can create billing system for super market. It can be used to reduce the difficulties to calculate large amount of money and used to reduce the error that can bo made by the billing machine. In this project I use python language, because it is very effective to use and very simple to make code.